Refactor fizzbuzz code into node modules
