--- index.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>The FizzBuzz Generator</title>
-    <meta name="description" content="A simple app that generates the classic FizzBuzz solutions">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-  </head>
-  <body>
-    <main>
-      <section>
-        <h1>The FizzBuzz Generator</h1>
-        <form id="user-input">
-          <label for="start-value">Start value</label>
-          <input id="start-value" type="number" name="start">
-          <label for="end-value">End value</label>
-          <input id="end-value" type="number" name="end">
-          <input id="submit-button" type="submit" value="Submit">
-        </form>
-        <table id="results">
-          <h2>Results</h2>
-        </table>
-      </section>
-    </main>
-    <script src="js/fizzbuzz.js"></script>
-    <script src="js/main.js"></script>
-  </body>
-</html>
