+++ module/index.js
@@ -1,0 +1,10 @@
+"use strict";
+
+(function() {
+
+  var FizzBuzz = require("./constructor.js");
+  var _generateOutput = require("./generator.js");
+
+  module.exports = FizzBuzz;
+
+})();
